Add compile option to disable floating point usage
This commit adds the option to disable floating point processing in
QCBOR. To disable, the QCBOR_DISABLE_FLOATING_POINT preprocessor macro
needs to be defined.

e.g.:

 $ make CMD_LINE="-DQCBOR_DISABLE_FLOATING_POINT"

This removes the capability (and the code) of decoding floating point
types. The type is still recognised, so a meaningful
QCBOR_ERR_FLOAT_DISABLED error is returned when a floating point value
is encountered in a decoded qcbor. From the encoding interface the
floating point encoding functions are removed.