Switch to OpenSSL EVP API (#61)
This uses the OpenSSL EVP API for ECDSA and for hashing rather than some other older APIs.

The good thing about this is that it is up to date and is not using any to-be-deprecated OpensSSL APIs.  It works and is tested against OpenSSL 1.1.1 and 3.0. 

It will be easier to support other public signing algorithms with this. The hash integration layer is simpler and uses less stack.

The bad thing about this is that these APIs are oriented to DER-format signatures and that some extra code and memory is needed to convert to the simple concatenation format that COSE uses. There is also more malloc use. MbedTLS is much cleaner in this way.  But for most people that don't work in the embedded space, this is of little consequence. 


* Check point work towards using non-deprecated OSSL APIs

* working with the EVP API, but needs more clean up

* more fixes, getting close to done

* Near complete with updated OpenSSL support

* New OpenSSL integration is pretty much finished

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>