New sign API supporting COSE_Sign (#79)
This is t_cose_sign (not t_cose_sign1 which it is to eventually replace). It supports both COSE_Sign and COSE_Sign1.

It makes use of the new parameter handling functions so it supports custom parameters. This is in t_cose_parameters.h and is now public. 

This makes use of the abstract base class for COSE_Signature. This is in t_cose_signature_sign. 

There is one concrete instkantation of t_cose_signature_sign which is t_cose_signature_sign_ecdsa.

No work was done on the verification side yet.

The old t_cose_sign1 is still present.

* Add in the new parameter handling code

* New tbs_hash function; makefile fixes

* couple more makefile fixes

* Add API design; some parameter fixes

* First version of COSE_Sign + COSE_Sign1

* Seems to be working...

* Fix erroneous include

* add inline declaration to fix warning

* Fix return type for a few tests

* Add new source files to cmake

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>