Allow GetOutPlace to work with NULL buffers.
Previously, when used with a NULL buffer at a non-zero offset, the
UsefulOutBuf_GetOutBuf would add the offset to the null pointer and
return it to the caller. The caller would see a non-NULL buffer, but any
attempt to write to it would obviously fail.

This is fixed by returning a NULL buffer if the argument itself is NULL.