Replace t_cose_sign1 with compatibility layer (#89)
The t_cose_sign1 implementation now rests on the t_cose_sign implementation.  This PR also fixes a bunch of bugs in t_cose_sign to the point that it mostly works properly for COSE_SIgn1. It doesn't work properly for COSE_Sign yet.

This PR fixes a bunch of bugs in header parameter encoding and decoding, something that Mac and Encrypt will make use of.

Also, some tests are not passing. They are disabled.  They will be fixed on a more incremental pace.


* t_cose_sign1 build on top of t_cose_sign

* signing compatibility layer is passing most tests

* compatibility mode tests passing; dup encode param detection

* Checkpoint COSE_Sign1 compatibility layer

* checkpoint compatibiliy work -- tests are passing

* add files to CMakeLists

* Check point work on option flags

* tdv tests passing

* Fixes for parameter look up functions

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>