Merge master with RSA and EdDSA into dev (#121)
Brings in support for RSA and EdDSA signing from the main branch.

Rename the ECDSA signer/verifier to "main" as it handles RSA and ECDSA. EdDSA is in a separate signer/verifier.

Short-circuit is no longer a fake mode of ECDSA keying off a special key ID or a signer/verifier of its own. It is now what the test crypto uses for signing. It has its own algorithm ID and is invoked by algorithm ID. It is supported by the "main" signer, but only when test crypto is used.

Many test cases were upgraded so they run with real crypto instead of only with short-circuit signatures. This increases test coverage when test crypto is used.

Interface change to signer/verifiers so that the full option flag set is passed to them.

This is a fairly large PR that touches a lot of files.


* Fix include directory ordering in Makefile.test

* Add support for RSASSA-PSS signatures. (#90)

This is described in RFC8230. Both OpenSSL and MbedTLS support is added.

The test suite includes some sign/verify self-tests, along with some
"known good" signatures, that are verified. As additional confirmation,
I've made sure the signatures produced by t_cose could be verified by
the pycose library, and vice-versa.

Reduces stack requirement for ECDSA signature by making use of OpenBytes API in QCBOR 1.1.

* Address CR comments.

* Fix formatting of known good signatures.

* Fix RSASSA-PSS warnings

* Fix warnings

* Fix some makefile and test.

* Fix more warnings

* Add pragmas to ignore Wcast-qual on OpenSSL calls.

* Remove extraneous includes.

* Add check for QCBOR1.1

* Update README

Co-authored-by: Paul Liétar <lietarpaul@microsoft.com>

* Add Paul Liétar  to credits

* Fix rarely occuring memory leak (#94)

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>

* Add EdDSA signatures support. (#92)

Support for EdDSA signature is added for PSA and OpenSSL crypto.

Thanks Paul!

* A few straggler nits from EdDSA addition (#98)

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>

* checkpoint -- most tests are passing now

* Fix the cmake file

* All tests passing, but still work to do....

* All tests passing

* Short circuit signer now part of the main signer

* Remove short from cmake

* Another cmake fix

* Tody cwt test; will it fix CI issue?

* Cmake fixes for CI

* rename ecdsa signer/verifier to main

* Fix installation of eddsa related headers

* More minor tweaks and fixes

* fix compiler warnings; add signature header for eddsa

* documentation and formatting

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>
Co-authored-by: Paul Liétar <plietar@users.noreply.github.com>
Co-authored-by: Paul Liétar <lietarpaul@microsoft.com>