Fix static analyzer warnings for adding empty data/strings (#170)
CBOR allows empty text and byte strings and QBOR supports them with AddText() and AddBytes() with a size of zero.

This results in a call to memmove with a NULL pointer. It was always with a zero length. Implementations of memmove tolerate this, but it is not recommended and static analyzers complain.

This PR fixes this. It also adds tests for this condition. It also documents that adding strings of zero length is supported in UsefulBuf.

This addresses #164



* Correct support for adding empty text/byte strings

* Update change log

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>