Fix static analyzer complaint about INT64_MIN for big floats and decimal fractions (#173)
This fixes potential static analyzer complaints about INT64_MIN when decoding big floats and decimal fraction tags. This is only for the value INT64_MIN (which is a special case for two's complement arithmetic). 

Both gcc and clang compilers seem to handle INT64_MIN min correctly so it is likely there is no problem for most uses.

A test for this has been added.

The inline designator has been removed from some of the functions to let the compiler/optimizer figure out whether inline is better or not.

Lots of comments have been added to the big float and decimal fraction decoding. 




* Improve code for INT64_MIN; better comments

* comment improvements

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>