Better AEAD crypto adaptor; HPKE uses AEAD adaptor. (#142)
The crypto adaptor layer for AEAD is improved.

Some small changes in the encrypt and decrypt t_cose public APIs were made to bring it more in line with t_cose conventions.

HPKE is now calling AEAD through the crypto adaptor layer rather than directly calling PSA APIs. (Still lots of work to move the rest of HPKE to crypto adaptors).

There is an OpenSSL implementation of the AEAD crypto adaptor layer. This was done to validate the design of the AEAD crypto adaptor layer.

Started a new set of tests for the crypto adaptor layer. This was motivated because the AEAD layer for OpenSSL is very complex and will be difficult to test through the public t_cose APIs that indirectly call AEAD. One of the difficulties with OpenSSL AEAD is that it allocated memory. Another is that it uses int rather that size_t for lengths. Another is that the APIs aren't well documented. Another is that they don't check lengths.

These new set of tests can be used to test other parts of the crypto adaptor layer. These tests ARE independent of the crypto library which is very good. They are not against public APIs though so the can't be against public t_cose APIs, such as against an installed binary library.

Also fixes some unrelated warnings from running tdv/b.sh.



* Support 'Make warn'

* Check point work on AEAD layer

* add simple test for aead; aead for openssl

* Fix cmake compilation

* Fake AEAD implementation for test crypto adaptor

* More testing is working. Mess with IV len in OpenSSL

* Tests are running; tidying up...

* Fix a bunch of warnings

* Fix warning for openssl and compilation error with EDDSA disabled

* Error codes and loose ends

* import symmetric cleaned up some

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>