Optimize algorithm ID and error code mapping (#145)
An int16 to int16 mapping function is added for general purpose mapping of error codes. Also works for algorithm IDs in some cases. The main purpose of this is to reduce object code.

The algorithm support check function is also better organized and there is less clutter in the crypto adapter interface.

A lot of comparisons of object code size were made creating this. Turns out case statements that map one integer to another are quite efficient. gcc and clang know how to optimize them very well.

The main point of this PR is code size reduction and tidiness. There is no semantic change.


* Support 'Make warn'

* utils getting done

* Int maps

* Fix mapping of short circuit algorithm IDs

* Clean up leftover junk; fix warning when short circuit is off

* Convert hmac errors to int16 map

* Optimize main verifier algorithm list

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>