Big refactor making encrypt recipients an object, plus more...
The public API for COSE_Encrypt and COSE_Encrypt0 now accepts an "object" that implements COSE_Recipient.

This fixes a bug where the layer of CBOR array that holds multiple COSE_Recipients was missing, both for encrypt and decrypt.

This provides some large refactoring of the encryption examples. One reason is because the public API for encryption changed a lot. Another is to make the examples more useful to people that want a clear and simple example of each usage mode of encryption. There is some more work to do to this end.

Make most symmetric key operations take a t_cose_key as input. Eliminate all but one export of t_cose_key into bytes outside of the crypto adaptor layer so t_cose can work better with HSMs.

Key wrap and direct encryption work for OpenSSL!

OpenSSL crypto adaptor returns random numbers.

* Support 'Make warn'

* Getting started on full recipient object

* direct encryption working for OpenSSL, key wrap not quite

* lots of kw rework

* OSSL random number adaptor API works now

* Make system for encrypt, other odds and ends

* Straggler warnings and make issues

* More straggler include issues

* Fix cmake file

* More stragglers for OSSL

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>