Verification of COSE_Sign with multiple COSE_Signature(s) (#153)
This mostly completes work on t_cose_sign_verify() so that it can verify a COSE_Sign with multiple signatures with multiple verifiers with multiple algorithms.

To actually verify multiple signatures with multiple verifiers, an updated QCBOR is needed. Multiple signatures with one verifier or one signature with multiple verifiers works with old versions of QCBOR. The new version of QCBOR is detected by #ifdef.

Kid matching is added to the main and eddy verifiers.

A base class for all COSE_Signature and COSE_Recipient objects now exists. It allows the linked list code to be shared across all.

* sign & verify of COSE_Sign with two signatures works

* Fix to last merge-commit

* verify all signatures works, returning parameters properly

* stragglers to make tests run correctly

* Minor fixes for the full test fan out and examples

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>