Cose_Recipient decryption object (#154)
This makes the implementations of decrypting a COSE_Recipient into an object (like signers, encryptors...)

Moves the HPKE COSE_Recipient decryption into it's own file so now COSE_Encrypt and COSE_Encrypt0 implementations are clean and don't do any COSE_Recipient stuff like HPKE.

Decoding multiple COSE_Recipients are still not supported (but the object structure will support them).

Clarify that direct encryption is NOT the same as Encrypt0 (9052 is sloppy on this, but in the end it is clear that they are not the same). Direct encryption requires a COSE_Recipient. Encrypt0 never has a COSE_Recipient.

The option_flag to specify detached is removed. The design for determining COSE_Encrypt vs COSE_Encrypt0 is changed.

Support for kid in the main body of COSE_Encrypt and COSE_Encrypt0 is removed as it is discouraged by 9052. It will be put back when general support for parameters is added. COSE_Recipients do support kid.

Setting the decryption keys is per (COSE_Recipient) t_cose_recipeint_dec instance now. You can set a key on on the the Encrypt and Encrypt0 bodies, but it is only the CEK and it is only required for Encrypt0.


* First check-in of decryption object work; not complete...

* Keywrap COSE_Recipient is functioning

* Empty protected is now 0 length bstr, + lots...

* most tests passing

* restore parameters.c to as it was when this was branched

* restore sign-verify test

* restore param_tests

* remove extraneous files

* hpke is working for psa

* Compiler warnings from tdv #define fan out

* encryption example stragglers

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>