Add Makefile.common and other Makefile improvements
Makefile.common now contains the main dependencies for the t_cose library and test suite. These do not vary by crypto adapter.

The Makefiles are cleaned up some and made more similar to each other.

Install of the public headers is corrected and complete.

The #includes for all the source files are ordered more clearly. There is still some dependency on nested #includes in .h files, but it is minimized.

Groups are made in the Xcode project file for the different COSE objects.


* Progress on header file inclusion and dependency

* more makefile improvements

* Update OSSL makefile

* Modernize Makefile.test; all tests passing for ossl

* Fix more dependencies; organize XCode project

* tidy up #includes

* Organize inc files in XCode

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>