restore inc/qcbor.h for backwards compatibility (#181)
A long time ago there was just inc/qcbor.h. This file was moved to inc/qcbor/qcbor.h to facilitate installation in /usr/local/include. inc/qcbor.h was retained for a while, but got lost in a big merge in 2020. This PR restores it. It simply includes inc/qcbor/qcbor.h.