Add restartable api, building on crypto context (#181)
* Remove T_COSE_CRYPTO_MAX_HASH_SIZE

T_COSE_CRYPTO_MAX_HASH_SIZE is not relevant anymore as all signers can
define their respective max value.

Change-Id: I532a97baf9bc2a1952ff6d8ab7c233ab95796b7a

* Add restartable signing option to t_cose

This commit adds the option to execute the signing of a COSE object in
a restartable manner. This means that the function call executing the
signing may return before the signing of the COSE object is complete
with T_COSE_ERR_SIG_IN_PROGRESS. In this case the function needs to be
called again with the same parameters, until T_COSE_SUCCESS or an error
is returned. Restartable option is enabled to ECDSA, but not EdDSA.

This commit also updates the PSA, OpenSSL and test crypto adapters so
that they return error when called in restartable manner.

Change-Id: Ifdb9145d6c538ffe1dd533d626e13e11c37cd64a

* Add restartable mode to PSA Crypto adapter

Change-Id: I8083f0cad83396075131a8df476909a58edbd5b6

* Add restartable implementation to test

* Implement restartable behaviour in the test adapter
* Add testcases that exercise the restartable implementation

Change-Id: I002b72cfd0cfd50ec0bf21483810964f62233a79