Not-quite-running esdh decrypt (#238)
This adds an ESDH COSE_Recipient for decryption. It is most of the code needed except the info struct and mostly runs but it probably doesn't work with actual keys and actual encrypted messages.

Possibly something is wrong with the crypto layer for ECDH or the key initialization in the tests and examples.

So far I haven't found any samples to test against. Jim's samples all use string labels. They are not supported here yet so they can't be used. The encrypt functionality here is not working either.


* First, non-running, esdh decrypt

* Lots of progress on ESDH decrypt including EC2 key decode

* A few bug fixes and a test that doesn't work yet

* Fix test build and warning

* Fix warning

* int_fast32_t -> int32_t

* back out test to see if it fixes CI errors

* Back out more stuff to debug CI

* Re add to debug CI

* try again

* re add more

* Add more back in

* Add more back in

* Back out t_cose_crypto changes as they are not really part of this PR

* Add more back in

* Clean up test a little

* Improve naming. Fix warnings

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>