KDF context tests passing; fix big memory leak (#258)
This completes work on the KDF context, PartyU, PartyV and such.  It also adds support for the salt. There's a big recommendation for what to do with the KDF context.

This also fixes a big leak of t_cose_keys for encryption.

This also tidy's up the encryption code some.



* KDF context tests mostly done and passing

* Documentation improvement

* Fix big memory/key leak; tests now running

* Fix minor test stuff caught in CI

* A few documentation tweaks

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>