Rework so COSE_Signature implementations are simpler + easier DECODE_ONLY (#257)
This is a major change to the internal way COSE_Sign is processed. No big API changes.

Verifier objects no longer need to decode COSE_Signatures and are half as complex as before. The COSE_SIgnature decoding is centralized. This also allows DECODE_ONLY with no verifiers (as long as COSE_Signatures don't use non-string or integer headers or need buffer size calculations).

Tests for DECODE_ONLY with multiple signatures




* Might fix aux buf size calc for multiple signatures

* Rework multiple COSE_Signatures

* Remove files that shouldn't have been added

* Documentation and clean up

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>