Correct doc & tests for mixed traversal in bounded maps (#191)
This does not change the behavior of QCBOR.

The changes the documentation to match the behavior. It also tests the behavior to make
sure it is correct.

The behavior at issue is where the traversal cursor ends up after fetching items by label in a map that has been entered. The retained behavior from QCBOR 1.2 is that fetching a non-aggregate does not affect the traversal cursor, but entering an aggregate sub map by label and exiting it does affect the traversal cursor.

The behavior is retained because it is usable, a change would bring backwards compatibility issues and the fix requires an increase in decode context size and in code complexity.


See also #149

* Correct doc & tests for mixed traversal in bounded maps

* straggler conflict marker

* Documentation updates

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>