Add user-defined/set errors; Fix minor VGetNext error handling bug
QCBORDecode_SetError() is added so QCBOR users can set an error and have it propagated up.

The data and label type of the item returned by VGetNext is correctly set to NONE when the error occurs right away, but was not set to NONE when the error occurred previously. This is fixed. This is a minor bug and only affected error handling use that relied solely on checking that the data and label were NONE.

Added more testing of error processing.


* Add SetError(); VGetNext error bug fix

* Documentation fixes

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>