Defensive handling of reserved values in additional info (#249)
This restores the defensive handling of reserved values (28, 29, 30) in additional info in the CBOR head that was present in QCBOR 1.3.

There are no problems or test failures in 1.4, but this fix should be picked up.

The change is to put the QCBORItem initialization back in the right place. From analysis of the code, the only effect was when the header decode returned QCBOR_ERR_UNSUPPORTED, a recoverable error. It is assumed that no caller would examine the secondary fields of QCBORItem when this error occurs, so there's no issue.

* Update version indicators to 1.4

* initialization fix

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>