Spiffy decode GetString() now ignores tag numbers (#260)
QCBORDecode_GetTextString() and QCBORDecode_GetByteString() now do not error out if there are tag numbers on the strings being decoded. This is to be consistent with all the other GetXXXX functions that also do not error out.

This is a behavior change, but it is assumed that no one relies on this behavior in QCBOR because most GetXXXX functions don't do it. Checking for extraneous tags is done with QCBORDecode_GetNthTagOfLast(). This works for of the GetXxxx functions and is what should be to replace reliance old QCBORDecode_GetByteString() behavior.


* Spiffy decode GetString() now ignores extra tag numbers

* fix use when tags are disabled

* improve tests

* fix #ifdef fan out

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>