Introduce T_COSE_OPT_ENABLE_NON_AEAD (#284)
This PR mitigates AEAD to non-AEAD downgrade attacks by warning the library uses to comply with RFC 9459.
Non AEAD algorithms are disabled by default, and will be enabled only when the T_COSE_OPT_ENABLE_NON_AEAD option is specified by the library caller.

The library callers using AES-CTR or AES-CBC will lose the backward compatibility, but it is crucial avoiding them to be vulnerable.





* update: disable non AEAD algorithms by default and introduce T_COSE_OPT_ENABLE_NON_AEAD to avoid unintentional use of them

* update: warn non AEAD use on T_COSE_OPT_ENABLE_NON_AEAD

* update: return T_COSE_ERR_NON_AEAD_DISABLED in both encrypt and decrypt functions

* fix: error code to T_COSE_ERR_NON_AEAD_DISABLED

* update: the usage of T_COSE_OPT_ENABLE_NON_AEAD

---------

Co-authored-by: Laurence Lundblade <laurencelundblade@users.noreply.github.com>