NaN payload bug fix; restructure some of the floating-point processing and disabling
This fixes a bug NaN payload preferred serialization where silent NaNs with payloads were incorrectly converted to quiet NaNs on some CPUs. Internal single-to-double conversions are now implemented in SW rather than using CPU.

Previously single-to-double conversion was disabled with QCBOR_DISABLE_FLOAT_HW_USE. Now it is disabled with QCBOR_DISABLE_PREFERRED_FLOAT. This is a behavior change and the reason the minor version number is incremented.

This also fixes a few error codes related to disabled floating-point features.

This updates the documentation on floating-point feature disabling.


* Single to Double conversion now uses SW, not CPU

* Fix #ifdef fan out and documentation

* Fix some error numbers; code tidy

* Fix error code

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>