Complete implementation and test of v2 tag decoding
This completes implementation and test of the v2 tag decoding design. It is an important milestone.

There was little change in the API, some error codes were rearranged and QCBORDecode_GetNthTagNumber renamed to QCBORDecode_NthTagNumber, but much of the internal implementation of the spiffy decode of QCBOR-supported tags was rearranged.


* Checkpoint work to complete tag decoding

* Expanded tag decode test and they pass

* getting very close to done

* Documentation fixes

* Clean up and TODOs

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>