Fix entering an empty bstr at end of input (#337)
QCBORDecode_EnterBstrWrapped() would fail if called on an empty byte string that occurred at the end of the CBOR input. It's a somewhat degenerate and very unlikely scenario, but it should not fail. Trying to fetching anything from inside this wrapped CBOR will of course correctly fail.

Test coverage for this and a few related conditions is added.

* Fix/test entering an empty bstr at end of input

* More code tidy up and optimization

---------

Co-authored-by: Laurence Lundblade <lgl@securitytheory.com>